Projects/Unsupervised_PTJC/app/requirements.txt
nibabel
torch
torchvision
monai
pandas
numpy
Pillow
scikit-image
scikit-learn
streamlit
matplotlib
opencv-python